pyproject.toml
[project]
name = "streamlit-gst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb==1.5.5",
    "ipykernel==7.2.0",
    "langchain==1.2.14",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.24",
    "langchain-groq==1.1.2",
    "langchain-openai==1.1.12",
    "pydantic==2.12.5",
    "pymupdf==1.27.2.2",
    "python-dotenv==1.2.2",
    "sentence-transformers==5.3.0",
    "fastapi==0.109.0",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "uuid==1.30",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiosignal==1.4.0",
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "appnope==0.1.4",
    "asttokens==3.0.1",
    "attrs==26.1.0",
    "bcrypt==5.0.0",
    "blinker==1.9.0",
    "build==1.4.2",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.20",
    "decorator==5.2.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "executing==2.2.1",
    "filelock==3.25.2",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "googleapis-common-protos==1.73.1",
    "groq==0.37.1",
    "grpcio==1.80.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.8.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "ipython==9.12.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "kubernetes==35.0.0",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.1.4",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.24",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.1",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.18.1",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "numpy==2.4.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.24.4",
    "openai==2.30.0",
    "opentelemetry-api==1.40.0",
    "opentelemetry-exporter-otlp-proto-common==1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.40.0",
    "opentelemetry-proto==1.40.0",
    "opentelemetry-sdk==1.40.0",
    "opentelemetry-semantic-conventions==0.61b0",
    "orjson==3.11.8",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "pandas==3.0.2",
    "parso==0.8.6",
    "pexpect==4.9.0",
    "pillow==12.2.0",
    "platformdirs==4.9.4",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.6",
    "psutil==7.2.2",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==23.0.1",
    "pybase64==1.4.3",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pydeck==0.9.1",
    "pygments==2.20.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "regex==2026.3.32",
    "requests==2.33.1",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.3",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.48",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "toml==0.10.2",
    "tornado==6.5.5",
    "tqdm==4.67.3",
    "traitlets==5.14.3",
    "transformers==5.4.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.42.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "wcwidth==0.6.0",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
    "python-multipart>=0.0.22",
]
